+++ eslint-configs/eslint-config-seekingalpha-react/rules/eslint-plugin-jest/index.js
@@ -34,6 +34,63 @@ module.exports = {
 
     // https://github.com/jest-community/eslint-plugin-jest/blob/HEAD/docs/rules/no-truthy-falsy.md
     'jest/no-truthy-falsy': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/consistent-test-it.md
+    'jest/consistent-test-it': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/expect-expect.md
+    'jest/expect-expect': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/lowercase-name.md
+    'jest/lowercase-name': 'off',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-jasmine-globals.md
+    'jest/no-jasmine-globals': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-jest-import.md
+    'jest/no-jest-import': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-large-snapshots.md
+    'jest/no-large-snapshots': [
+      'error',
+      {
+        maxSize: 50,
+      },
+    ],
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-test-prefixes.md
+    'jest/no-test-prefixes': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-test-return-statement.md
+    'jest/no-test-return-statement': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/prefer-expect-assertions.md
+    'jest/prefer-expect-assertions': 'off',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/prefer-inline-snapshots.md
+    'jest/prefer-inline-snapshots': 'off',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/prefer-strict-equal.md
+    'jest/prefer-strict-equal': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/prefer-to-be-null.md
+    'jest/prefer-to-be-null': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/no-hooks.md
+    'jest/no-hooks': 'off',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/prefer-to-be-undefined.md
+    'jest/prefer-to-be-undefined': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/valid-describe.md
+    'jest/valid-describe': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/valid-expect.md
+    'jest/valid-expect': 'error',
+
+    // https://github.com/jest-community/eslint-plugin-jest/blob/master/docs/rules/valid-expect-in-promise.md
+    'jest/valid-expect-in-promise': 'error',
+
   },
 
 };