+++ eslint-configs/eslint-config-seekingalpha-react/rules/eslint-plugin-react/jsx.js
@@ -217,6 +217,9 @@ module.exports = {
       },
     ],
 
+    // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-danger.md
+    'react/no-danger': 'error',
+
   },
 
 };